bioRefProject/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
black==22.1.0
bleach==4.1.0
blinker==1.4
Bottleneck @ file:///C:/ci/bottleneck_1607557040328/work
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
datasets==1.18.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
Django==4.0.2
djangorestframework==3.13.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.2
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.1.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.1
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numexpr @ file:///C:/ci/numexpr_1640704337920/work
numpy @ file:///C:/ci_310/numpy_and_numpy_base_1643798589088/work
oauthlib==3.2.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1644984307868/work
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentencepiece==0.1.95
sentry-sdk==1.5.5
seqeval==1.2.2
shortuuid==1.0.8
simpletransformers==0.63.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.5.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tokenizers==0.11.5
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets==5.1.1
transformers==4.16.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
wandb==0.12.10
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wincertstore==0.2
xxhash==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
